blog_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pytest==8.3.3
httpx==0.27.2
ruff==0.6.5
pre-commit==3.8.0
